move pane hit test logic into separate module
